backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-nmap>=0.7.1
scapy>=2.5.0
manuf>=1.1.5
httpx>=0.27.0
apscheduler>=3.10.4
python-multipart>=0.0.9
netifaces>=0.11.0
websockets>=12.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
